Major Changes. Currently client is unreliable. Randomly disconnects from game. Occasionally there are stray packets that are picked up as the wrong type, causing some information to be incorrect. Under some circumstances, the client WILL stay connected for long periods of time. It seems TNT and mass block changes cause it to become unstable.
In addition to the problems, I have added a PlayerLocationChanged event. As far as I know, this is (probably most of the time) accurate. More events to come, including PlayerRotationChanged (though, that seems like it wouldn't be very useful). 

The major changes earlier mentioned are that most of the packet types parse the correct amount of data, thus leaving very few stray packets. However, these do occur.

Finally, I have also added SetPlayerLocation(), but it currently does nothing because the client doesn't actually send its location to the server yet. This is mainly because I currently do not know how often it should be sent, or what triggers it to be sent. I will probably just send the location data when the client receives it from the server.